refactor: extract user register logic to UserRegisterService from UserService
