+++ src/main/java/com/depromeet/archive/api/user/RegisterController.java
@@ -1,7 +1,7 @@
 package com.depromeet.archive.api.user;
 
 import com.depromeet.archive.api.dto.user.OAuthRegisterDto;
-import com.depromeet.archive.domain.user.UserService;
+import com.depromeet.archive.domain.user.UserRegisterService;
 import com.depromeet.archive.domain.user.command.PasswordRegisterCommand;
 import com.depromeet.archive.domain.user.info.UserInfo;
 import com.depromeet.archive.infra.user.oauth.OAuthUserService;
@@ -25,7 +25,7 @@ import javax.validation.Valid;
 @RequiredArgsConstructor
 public class RegisterController {
 
-    private final UserService userService;
+    private final UserRegisterService userRegisterService;
     private final OAuthUserService oAuthUserService;
     private final PasswordEncoder encoder;
 
@@ -36,7 +36,7 @@ public class RegisterController {
     @PostMapping("/register")
     public ResponseEntity<Void> registerUser(@RequestBody @Valid PasswordRegisterCommand command) {
         encryptPassword(command);
-        userService.registerUser(command);
+        userRegisterService.registerUser(command);
         return ResponseEntity.ok().build();
     }
 
@@ -45,7 +45,7 @@ public class RegisterController {
                                                           @RequestParam("provider") String provider,
                                                           @RequestBody(required = false) OAuthRegisterDto oAuthRegisterDto) {
         var oAuthRegisterInfo = oAuthUserService.getOAuthRegisterInfo(provider, oAuthRegisterDto);
-        var userInfo = userService.getOrRegisterUserReturnInfo(oAuthRegisterInfo);
+        var userInfo = userRegisterService.getOrRegisterUserReturnInfo(oAuthRegisterInfo);
         injectJwtToken(httpServletResponse, userInfo);
         return ResponseEntity.ok().build();
