+++ src/main/java/com/depromeet/archive/config/AuthConfig.java
@@ -2,14 +2,14 @@ package com.depromeet.archive.config;
 
 import com.depromeet.archive.domain.archive.ArchiveRepository;
 import com.depromeet.archive.domain.user.UserLoginService;
-import com.depromeet.archive.domain.user.UserService;
+import com.depromeet.archive.domain.user.UserRegisterService;
 import com.depromeet.archive.security.authorization.permissionhandler.ArchiveAdminOrAuthorChecker;
 import com.depromeet.archive.security.general.UserNamePasswordAuthenticationProvider;
+import com.depromeet.archive.security.oauth.OAuthUserService;
 import com.depromeet.archive.security.result.LoginFailureHandler;
 import com.depromeet.archive.security.result.LoginSuccessHandler;
 import com.depromeet.archive.security.token.HttpAuthTokenSupport;
 import com.depromeet.archive.security.token.TokenProvider;
-import com.depromeet.archive.security.oauth.OAuthUserService;
 import com.depromeet.archive.security.token.jwt.JwtTokenProvider;
 import com.depromeet.archive.security.token.jwt.JwtTokenSupport;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -32,9 +32,8 @@ public class AuthConfig {
     }
 
     @Bean
-    public OAuthUserService oAuthUserService(UserService userService) {
-        OAuthUserService oAuthUserService = new OAuthUserService(userService);
-        return oAuthUserService;
+    public OAuthUserService oAuthUserService(UserRegisterService userRegisterService) {
+        return new OAuthUserService(userRegisterService);
     }
 
     @Bean
@@ -53,7 +52,9 @@ public class AuthConfig {
     }
 
     @Bean
-    public LoginFailureHandler failureHandler() {return new LoginFailureHandler(); }
+    public LoginFailureHandler failureHandler() {
+        return new LoginFailureHandler();
+    }
 
     @Bean
     public PasswordEncoder encoder() {
